Remove hasCommentInside functions
