+++ detekt-psi-utils/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/Junk.kt
@@ -1,28 +1,11 @@
 package io.gitlab.arturbosch.detekt.rules
 
-import org.jetbrains.kotlin.com.intellij.openapi.util.Key
-import org.jetbrains.kotlin.com.intellij.psi.PsiComment
-import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.psi.KtBlockExpression
 import org.jetbrains.kotlin.psi.KtCallExpression
 import org.jetbrains.kotlin.psi.KtClass
-import org.jetbrains.kotlin.psi.KtClassOrObject
 import org.jetbrains.kotlin.psi.KtQualifiedExpression
-import org.jetbrains.kotlin.psi.KtTreeVisitorVoid
 import org.jetbrains.kotlin.resolve.BindingContext
 import org.jetbrains.kotlin.resolve.bindingContextUtil.isUsedAsExpression
-
-fun KtClassOrObject.hasCommentInside() = this.body?.hasCommentInside() ?: false
-
-fun PsiElement.hasCommentInside(): Boolean {
-    val commentKey = Key<Boolean>("comment")
-    this.acceptChildren(object : KtTreeVisitorVoid() {
-        override fun visitComment(comment: PsiComment) {
-            putUserData(commentKey, true)
-        }
-    })
-    return getUserData(commentKey) == true
-}
 
 fun KtClass.companionObject() = this.companionObjects.singleOrNull { it.isCompanion() }
